fixes example project crashes in iPad
