+++ Example/PagingMenuControllerDemo2/ViewController1.swift
@@ -21,18 +21,5 @@ class ViewController1: UIViewController {
         textLabel.font = UIFont.systemFontOfSize(24)
         textLabel.text = "View Controller 1"
         view.addSubview(textLabel)
-        
-        let button = UIButton(type: UIButtonType.System)
-        button.setTitle("Change Menu", forState: UIControlState.Normal)
-        button.frame = CGRectMake(80, 300, 300, 30)
-        button.titleLabel!.textColor = UIColor.blackColor()
-        view.addSubview(button)
-        
-        button.addTarget(self, action: #selector(ViewController1.buttonTapped(_:)), forControlEvents: UIControlEvents.TouchUpInside)
-    }
-
-    func buttonTapped(sender: UIButton) {
-        self.menuItemDesc = "Changed"
-        self.updateMenuContent()
     }
 }
